+++ backend/internal/db/migrations/auth/000003_sessions.up.sql
@@ -2,7 +2,7 @@ create extension if not exists "pgcrypto";
 
 create table if not exists sessions (
     session_id uuid primary key default gen_random_uuid(),
-    user_id bigint not null,
+    user_id int not null,
     revoked boolean not null default false,
     revoke_at timestamptz,
     created_at timestamptz not null default now(),
