+++ backend/internal/db/migrations/chat/000002_conversation_members.up.sql
@@ -2,7 +2,7 @@ create type member_role as enum ('admin', 'member');
 
 create table if not exists conversation_members(
     conversation_id bigint not null,
-    user_id bigint not null,
+    user_id int not null,
     role member_role not null default 'member',
     joined_at timestamptz not null default now(),
