+++ backend/internal/models/auth.go
@@ -20,7 +20,7 @@ type GoogleUserInfo struct {
 type GoogleLoginResponse struct {
 	SessionId     uuid.UUID `json:"session_id"`
 	UUID          uuid.UUID `json:"uuid"`
-	UserId        int64     `json:"user_id"`
+	UserId        int32     `json:"user_id"`
 	ProfileExists bool      `json:"profile_exists"`
 }
 
@@ -35,7 +35,7 @@ type GoogleBirthdayResponse struct {
 }
 
 type SessionRedis struct {
-	UserID         int64 `json:"user_id"`
+	UserID         int32 `json:"user_id"`
 	SessionVersion int   `json:"session_version"`
 	Valid          bool  `json:"valid"`
 }
